+++ index.js
@@ -20,6 +20,8 @@ sp.on("dailymotion_check_delay", dailymotion_check_delay_onChange);
 let tabs = require("sdk/tabs");
 let windows = require("sdk/windows").browserWindows;
 
+let ContextMenu = require("sdk/context-menu");
+
 let {setInterval, setTimeout, clearInterval} = require("sdk/timers");
 
 function getPreferences(prefId){
@@ -31,6 +33,7 @@ function savePreference(prefId, value){
 }
 
 let myIconURL = self.data.url("live_offline_64.svg");
+let myIconURL_16 = self.data.url("live_offline_16.svg");
 
 let myIconURL_online = {
 	"16": "./live_online_16.svg",
@@ -237,6 +240,29 @@ function refreshStreamsFromPanel(){
 	setTimeout(waitToUpdatePanel, 5000);
 }
 
+let addStream_URLpatterns = {"dailymotion": [/^(?:http|https):\/\/games\.dailymotion\.com\/(?:live|video)\/([a-zA-Z0-9]+).*$/, /^(?:http|https):\/\/www\.dailymotion\.com\/(?:embed\/)?video\/([a-zA-Z0-9]+).*$/, /^(?:http|https):\/\/games\.dailymotion\.com\/[^\/]+\/v\/([a-zA-Z0-9]+).*$/],
+		"channel::dailymotion": [/^(?:http|https):\/\/(?:games\.|www\.)dailymotion\.com\/user\/([^\s\t\/]+).*$/, /^(?:http|https):\/\/(?:games\.|www\.)dailymotion\.com\/(?!user\/|monetization\/|archived\/|fr\/|en\/|legal\/|stream|rss)([^\s\t\/]+).*$/],
+		"hitbox": [/^(?:http|https):\/\/www\.hitbox\.tv\/(?:embedchat\/)?([^\/\?\&]+).*$/],
+		"twitch": [/^(?:http|https):\/\/www\.twitch\.tv\/([^\/\?\&]+).*$/,/^(?:http|https):\/\/player\.twitch\.tv\/\?channel\=([\w\-]+).*$/],
+		"beam": [/^(?:http|https):\/\/beam\.pro\/([^\/\?\&]+)/]
+	};
+let URLContext_Array = new Array();
+for(website in addStream_URLpatterns){
+	URLContext_Array = URLContext_Array.concat(addStream_URLpatterns[website]);
+}
+ContextMenu.Item({
+	label: _("Add this"),
+	image: self.data.url("../icon.png"),
+	context: [
+		ContextMenu.URLContext(URLContext_Array),
+		ContextMenu.SelectorContext("a[href]")
+	],
+	contentScriptFile: self.data.url("page_getUrlLink.js"),
+	onMessage: function(data){
+		console.info(`[ContextMenu] URL: ${data}`);
+		addStreamFromPanel({"ContextMenu_URL": data});
+	}
+});
 function display_id(id){
 	if(website_channel_id.test(id)){
 		return _("The channel %d", website_channel_id.exec(id)[1]);
@@ -245,31 +271,31 @@ function display_id(id){
 	}
 }
 let addStreamFromPanel_pageListener = new Array();
-function addStreamFromPanel(embed_list){
+function addStreamFromPanel(data){
 	let current_tab = tabs.activeTab;
 	let active_tab_url = current_tab.url;
-	console.info("Current active tab: " + active_tab_url);
-	let active_tab_title = current_tab.title;
+	
 	let type;
-	let patterns = {"dailymotion": [/^(?:http|https):\/\/games\.dailymotion\.com\/(?:live|video)\/([a-zA-Z0-9]*).*$/, /^(?:http|https):\/\/www\.dailymotion\.com\/(?:embed\/)?video\/([a-zA-Z0-9]*).*$/, /^(?:http|https):\/\/games\.dailymotion\.com\/[^\/]+\/v\/([a-zA-Z0-9]*).*$/],
-					"channel::dailymotion": [/^(?:http|https):\/\/(?:games\.|www\.)dailymotion\.com\/([a-zA-Z0-9\-_]*).*$/],
-					"hitbox": [/^(?:http|https):\/\/www\.hitbox\.tv\/(?:embedchat\/)?([^\/\?\&]*).*$/],
-					"twitch": [/^(?:http|https):\/\/www\.twitch\.tv\/([^\/\?\&]*).*$/,/^(?:http|https):\/\/player\.twitch\.tv\/\?channel\=([\w\-]*).*$/],
-					"beam": [/^(?:http|https):\/\/beam\.pro\/([^\/\?\&]*)/]};
 	let url_list;
-	if(typeof embed_list == "object"){
-		console.log(`Embed list (${active_tab_url})`);
-		console.dir(embed_list);
-		url_list = embed_list;
-		for(i of addStreamFromPanel_pageListener){
-			i.port.removeListener("addStream", addStreamFromPanel);
+	if(typeof data == "object"){
+		console.dir(data);
+		if(data.hasOwnProperty("ContextMenu_URL")){
+			url_list = [data.ContextMenu_URL];
+			type = "ContextMenu";
+		} else if(data.hasOwnProperty("embed_list")){
+			console.log("[Live notifier] AddStream - Embed list");
+			url_list = data.embed_list;
+			for(i of addStreamFromPanel_pageListener){
+				i.port.removeListener("addStream", addStreamFromPanel);
+			}
+			type = "embed";
 		}
-		type = "embed";
 	} else {
+		console.info("Current active tab: " + active_tab_url);
 		url_list = [active_tab_url];
 	}
 	for(let url of url_list){
-		for(source_website in patterns){
+		for(source_website in addStream_URLpatterns){
 			let website = source_website;
 			if(website_channel_id.test(source_website)){
 				website = website_channel_id.exec(source_website)[1];
@@ -277,7 +303,7 @@ function addStreamFromPanel(embed_list){
 			
 			let streamListSetting = new streamListFromSetting(website);
 			let streamList = streamListSetting.objData;
-			for(let pattern of patterns[source_website]){
+			for(let pattern of addStream_URLpatterns[source_website]){
 				let id = "";
 				if(pattern.test(url)){
 					id = pattern.exec(url)[1];
@@ -298,15 +324,17 @@ function addStreamFromPanel(embed_list){
 							onComplete: function (response) {
 								let id = id_toChecked;
 								data = response.json;
+								
+								console.group()
+								console.info(`${website} - ${current_API.url}`);
 								console.dir(data);
+								console.groupEnd();
 								
 								if(isValidResponse(website, data) == false){
 									if(website == "dailymotion" && (website_channel_id.test(source_website) || data.mode == "vod")){
 										let username = (data.mode == "vod")? data["user.username"] : data.username;
 										let id_username = `channel::${username}`;
 										let id_owner = `channel::${(data.mode == "vod")? data.owner : data.id}`;
-										
-										console.warn(id_username + id_owner)
 										
 										// Use username (login) as channel id
 										id = id_username;
@@ -347,12 +375,14 @@ function addStreamFromPanel(embed_list){
 			}
 		}
 	}
-	if(typeof embed_list != "object"){
-		let page_port = current_tab.attach({
-			contentScriptFile: self.data.url("page_getEmbedList.js")
-		});
-		addStreamFromPanel_pageListener.push(page_port);
-		page_port.port.on("addStream", addStreamFromPanel);
+	if(typeof data != "object" && type != "ContextMenu"){
+		if(!data.hasOwnProperty("embed_list")){
+			let page_port = current_tab.attach({
+				contentScriptFile: self.data.url("page_getEmbedList.js")
+			});
+			addStreamFromPanel_pageListener.push(page_port);
+			page_port.port.on("addStream", addStreamFromPanel);
+		}
 	} else {
 		doNotif("Live Notifier", _("No supported stream detected in the current tab, so, nothing to add."));
 	}
@@ -842,6 +872,10 @@ function isValidResponse(website, data){
 			break;
 		case "hitbox":
 			if(data.error == "live"){
+				console.warn(`[${website}] Unable to get stream state (error detected).`);
+				return false;
+			}
+			if(data.error == true){
 				console.warn(`[${website}] Unable to get stream state (error detected).`);
 				return false;
 			}